Update function 'entropy' in 'metrics.R'
breaks calculation in the function 'entropy' has the potential to exclude points when supplied bin-width argument is greater than 1. For example when zmax = 20.38 and by = 5 the last break point will be 20 and the maximum point will be excluded, returning an error. The proposed change uses the same method as implemented for a similar calculation in the function 'gap_fraction_profile'.

Former-commit-id: a6c54790218ee22bc47e45feee8f741f2be0d921
Former-commit-id: 6f68dad4f4b950df8a4b24cba60955965facb616
Former-commit-id: 438e2b281c5bc3675642b18cfe0c3160b80fe56e [formerly 544cbee7ccb14300e332f55ff3cf206062438045]
Former-commit-id: 04a7fac8118b6e20748e85ef45d10773f90f2ab3
Former-commit-id: 918a5abada555e569451872c650bc8170b62d4a8
Former-commit-id: 8d8bbd16463463d773cdc853b97253223e1d3ff0