bug in treetops ID check
There was a bug on treetop ID.
However, it raises a question about the expected third column, which seems to be an ID and not a Z value as it is given by tree_detection. Thus if several tree tops have the exact same height, the algorithm throws an error. And if several tops have the same integer part they will be identified with the same color. Thus
Thus, to make it work properly without the user-defined ID option (like in the other algorithms):
- I replaced treetops_df[, 3] by  treetops_df[[3]]
- I changed to generate a unique treetops_df integer ID systematically
If the possibility to have a user-defined treetop ID is wanted, it could be either given in a new argument or expressed specifically to be the third column or the RasterLayer value in the manual. In that case, this option should be thought for the other algorithms as well, to be homogeneous and to keep lastrees working, shouldn't it?