fixes bug where simple features with geometry column not named "geometry" would throw an error
Using sf::st_geometry() to access the geometry column of a simple feature ensures the geometry column (whatever it may be called) is used to get the geometry. If $geometry is used and the geometry column is called something different (e.g., lots of real-world instances where geometry column is called "geom") then NULL results, and sf::st_as_text() throws an error.